Physical-AI-Book/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
qdrant-client>=1.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai>=1.33.0
openai-agents
python-dotenv==1.0.0
tiktoken==0.5.2
pyyaml==6.0.1
pytest==7.4.3
httpx>=0.25.1
pytest-asyncio==0.21.1
slowapi==0.1.9
